A UniformControlParameter is provided.
This class allows one to replace algorithm control parameters with random
numbers, without the need to change any code in the algorithm.

For example, the scale parameter in DE can be replaced with a random
parameter for each trial vector to increase exploration.
